+++ effectful/internals/product_n.py
@@ -1,13 +1,33 @@
 import dataclasses
 import functools
-from collections.abc import Mapping
+from collections.abc import Callable, Mapping
+from typing import Any
 
 import tree
 
-from effectful.internals.runtime import CallByNeed
 from effectful.ops.semantics import apply, coproduct, handler
 from effectful.ops.syntax import defop
 from effectful.ops.types import Interpretation, Operation
+
+
+@dataclasses.dataclass
+class CallByNeed[**P, T]:
+    func: Callable[P, T]
+    args: Any  # P.args
+    kwargs: Any  # P.kwargs
+    value: T | None = None
+    initialized: bool = False
+
+    def __init__(self, func, *args, **kwargs):
+        self.func = func
+        self.args = args
+        self.kwargs = kwargs
+
+    def __call__(self):
+        if not self.initialized:
+            self.value = self.func(*self.args, **self.kwargs)
+            self.initialized = True
+        return self.value
 
 
 @defop
